requirements.txt
streamlit
requests
langchain
langchain-groq
langgraph
pydub
speechrecognition
audio-recorder-streamlit
python-dotenv
langchain-community
